Moved player location serialization to Board. Updated deserialization method to properly handle Board's occupancy counts.
